Improved adapting of icons/panels.
Before, icon and panel elements would be discarded based on the equality
of their node (meaning that, e.g. you could not change the title of a
node without causing its panel and icon elements to be destroyed). Now,
the check for whether an icon/panel element can be retained is done
based on the equality of templates.